Module Parse__Located

A parsing construct located within a file.

type 'a t

val make : Parse.Position.range -> 'a -> 'a t

make range object Make object located in range.

val value : 'a t -> 'a

value loc The located object.

val start : 'a t -> Parse.Position.t

start loc The start position of the located object loc.

val _end : 'a t -> Parse.Position.t

_end loc The end position of the located object loc.

val range : 'a t -> Parse.Position.range

range loc The pair consisting of start position and end position of the located object loc.