Module Parse__Parser_docu

The final parser.

A parser is a sink of token. As long as it signals needs_more p more token can be pushed into the parser via put token p or the input stream can be ended via put_end p.

If the parser does not need more token (has_ended p is equivalent to not (needs_more p)), then it has either succeeded or failed.

If it has succeeded the final value is available via final p.

There are two types of failure:

Syntax error: In that case failed_expectations p returns the list of failed expectations.

Semantic error: In that case failed_semantic p returns the encountered semantic error.