Module Std

Standard Data Types.

Common Interfaces

module Interfaces : sig ... end

Standard Module Types

Data Types

module Array : sig ... end

A thin wrapper around Stdlib.Array

module Deque : sig ... end

A double ended queue. Insert from both ends, pop only from the front.

module Int : sig ... end

Module to represent the type int

module List : sig ... end

A thin wrapper around Stdlib.List which avoids throwing exceptions and with some additional monadic functions.

module Result : sig ... end

Result: Handling results of operations which can fail

module String : sig ... end

A small wrapper around Stdlib.String with some extra functions.

module Unit : sig ... end

Module to represent the type unit.