Module type Std__Interfaces.SOURCE

Interface for a source of items i.e. a stream.

type item

Type of items coming from the source.

type t

t A stream of items.

val has_more : t -> bool

has_more s Does the stream s have more items?

val peek : t -> item

peek s Peek the next item from the stream s.

Precondition: has_more s

val advance : t -> t

advance s Pop the top item off the stream s.

Precondition: has_more s