Module Std__Result

Result: Handling results of operations which can fail

Overview

API

type ('a, 'e) t = ('a, 'e) Stdlib.result

'a is the result type in case of success and 'e' is the result type in case of failure. It is implemented by the ocaml type result from the ocaml standard library.

val return : 'a -> ('a, 'e) t

return a is the same as Ok a.

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

m >>= f

maps success result m to f a. In case of an error result f is not called and the error remains.

val let* : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

let* a = m in f a is the same as m >>= f

Monad

module Monad : functor (E : Std.Interfaces.ANY) -> sig ... end

The result type encapsulated in a module which satisfies the monadic interface.