Standard Types

Fmlib_std is a library of standard types. It has a lot of thin wrappers around modules of ocaml's standard library like Fmlib_std.Array, Fmlib_std.List etc. with some additional functionality and it contains some types which are not part of ocaml's standard library like e.g. Fmlib_std.Deque, Fmlib_std.Btree.

Usage

Install the library via

    opam install fmlib_std

In order to compile program named foo with the help of dune write a dune file like

    (executable
        (name foo)
        (libraries fmlib_std ...)
        ...
    )

API

Fmlib_std